refactor: split support log redaction
